now fetching details from the BatchRegistry contract && Separated the client logic.
