+++ packages/nextjs/utils/scaffold-eth/twMerge.ts
@@ -1,0 +1,6 @@
+import { ClassValue, clsx } from "clsx";
+import { twMerge } from "tailwind-merge";
+
+export default function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
